♻️(tariff) refactor tariff handling in API
