+++ src/api/tests/schemas/test_tariff_utils.py
@@ -212,4 +212,3 @@ def test_is_tariff_allowed_for_user(db_session):
     )
     allowed = UserFactory.create_sync(is_superuser=False, groups=[group])
     assert is_tariff_allowed_for_user(db_session, tariff.id, allowed) is True
-
