+++ src/client/qcc/endpoints/tariff.py
@@ -20,12 +20,20 @@ class Tariff(BaseEndpoint):
         from_: Optional[datetime] = None,
         to: Optional[datetime] = None,
         pdc: Optional[List[str]] = None,
+        current: Optional[bool] = None,
     ) -> AsyncIterator[dict]:
         """Query the /statique/tariff endpoint (GET)."""
         from_str = from_.isoformat() if from_ else None
         to_str = to.isoformat() if to else None
         params = dict(
-            p for p in (("from", from_str), ("to", to_str), ("pdc", pdc)) if p[1]
+            p
+            for p in (
+                ("from", from_str),
+                ("to", to_str),
+                ("pdc", pdc),
+                ("current", current),
+            )
+            if p[1] is not None
         )
 
         response = await self.client.get(f"{self.endpoint}/", params=params)
@@ -34,7 +42,7 @@ class Tariff(BaseEndpoint):
         except httpx.HTTPStatusError as err:
             raise APIRequestError(response.json()) from err
 
-        for tariff in response.json():
+        for tariff in response.json()["items"]:
             yield tariff
 
     async def applicable(
@@ -54,21 +62,3 @@ class Tariff(BaseEndpoint):
             raise APIRequestError(response.json()) from err
 
         return response.json()
-
-    async def associate(self, obj: dict) -> dict:
-        """Query the /statique/pointdechargetariff endpoint."""
-        response = await self.client.post("/statique/pointdechargetariff", json=obj)
-        try:
-            response.raise_for_status()
-        except httpx.HTTPStatusError as err:
-            raise APIRequestError(response.json()) from err
-
-        return response.json()
-
-    async def delete(self, id_: str) -> None:
-        """Query the /statique/tariff/{id_} endpoint (DELETE)."""
-        response = await self.client.delete(f"{self.endpoint}/{id_}")
-        try:
-            response.raise_for_status()
-        except httpx.HTTPStatusError as err:
-            raise APIRequestError(response.json()) from err
